requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
isort==5.13.2
black==24.8.0

